remove some duplicated code
