+++ packages/react-components/lib/map/waypoint-marker.tsx
@@ -2,7 +2,7 @@ import { makeStyles } from '@material-ui/core';
 import Debug from 'debug';
 import React from 'react';
 import * as RmfModels from 'rmf-models';
-import { fromRmfCoords } from '../geometry-utils';
+import { fromRmfCoords } from '../utils/geometry';
 
 const debug = Debug('Map:WaypointMarker');
